Move global methods into legacy
